Separate new from kernel trait and refactor filecoin kernel into separate file
